fixed leak

+++ srcb/ft_len.c
@@ -39,7 +39,7 @@ int ft_unsigned_len(unsigned int nb)
 	return (i);
 }
 
-int ft_hexa_x_len(unsigned int nb)
+int ft_hexa_x_len(unsigned long nb)
 {
 	int	i;
